+++ src/test/java/janggi/position/PositionTest.java
@@ -1,9 +1,8 @@
-package janggi;
+package janggi.position;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
-import janggi.position.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
